+++ src/Calamity/Commands/Utils.hs
@@ -5,7 +5,8 @@ module Calamity.Commands.Utils
     , buildCommands
     , buildContext
     , handleCommands
-    , findCommand ) where
+    , findCommand
+    , CmdInvokeFailReason(..) ) where
 
 import           Calamity.Cache.Eff
 import           Calamity.Metrics.Eff
@@ -45,32 +46,15 @@ mapLeft :: (e -> e') -> Either e a -> Either e' a
 mapLeft f (Left x)  = Left $ f x
 mapLeft _ (Right x) = Right x
 
-data FailReason
-  = NoCtx
-  | NF [L.Text]
-  | ERR Context CommandError
+data CmdInvokeFailReason
+  = NoContext
+  | NotFound [L.Text]
+  | CommandInvokeError Context CommandError
 
 -- | Construct commands and groups from a command DSL, then registers an event
 -- handler on the bot that manages running those commands.
 --
--- This registers the following event handler for: @"invoke-command" ('Message',
--- 'L.Text')@ that takes a message and the command string to invoke (without a
--- prefix)
---
--- Returns an action to remove the event handler, and the 'CommandHandler' that was constructed
-addCommands :: (BotC r, P.Member ParsePrefix r) => P.Sem (DSLState r) a -> P.Sem r (P.Sem r (), CommandHandler, a)
-addCommands m = do
-  (handler, res) <- buildCommands m
-  remove <- react @'MessageCreateEvt $ \msg -> do
-    parsePrefix msg >>= \case
-      Just (prefix, cmd) ->
-        handleCommands handler msg prefix cmd
-      Nothing -> pure ()
-  remove' <- react @('CustomEvt "invoke-command" (Message, L.Text)) $ \(msg, cmd) -> do
-    handleCommands handler msg "" cmd
-  pure (remove *> remove', handler, res)
-
--- | Manages parsing messages and handling commands for a CommandHandler.
+-- Returns an action to remove the event handler, and the 'CommandHandler' that was constructed.
 --
 -- ==== Custom Events
 --
@@ -80,7 +64,7 @@ addCommands m = do
 --
 --         Fired when a command returns an error.
 --
---     2. @"command-not-found" ['Data.Text.Lazy.Text']@
+--     2. @"command-not-found" ('Calamity.Types.Model.Channel.Message', '['Data.Text.Lazy.Text'])@
 --
 --         Fired when a valid prefix is used, but the command is not found.
 --
@@ -88,26 +72,39 @@ addCommands m = do
 --
 --         Fired when a command is successfully invoked.
 --
+addCommands :: (BotC r, P.Member ParsePrefix r) => P.Sem (DSLState r) a -> P.Sem r (P.Sem r (), CommandHandler, a)
+addCommands m = do
+  (handler, res) <- buildCommands m
+  remove <- react @'MessageCreateEvt $ \msg -> do
+    parsePrefix msg >>= \case
+      Just (prefix, cmd) -> do
+        r <- handleCommands handler msg prefix cmd
+        case r of
+          Left (CommandInvokeError ctx e) -> fire $ customEvt @"command-error" (ctx, e)
+          Left (NotFound path)            -> fire $ customEvt @"command-not-found" (msg, path)
+          Left NoContext                  -> pure () -- ignore if context couldn't be built
+          Right ctx        -> do
+            cmdInvoke <- registerCounter "commands_invoked" [("name", S.unwords $ commandPath (ctx ^. #command))]
+            void $ addCounter 1 cmdInvoke
+            fire $ customEvt @"command-invoked" ctx
+      Nothing -> pure ()
+  pure (remove, handler, res)
+
+-- | Manages parsing messages and handling commands for a CommandHandler.
+--
+-- Returns Right if the command succeeded in parsing and running, Left with the
+-- reason otherwise.
 handleCommands :: (BotC r, P.Member ParsePrefix r)
                => CommandHandler
                -> Message -- ^ The message that invoked the command
                -> L.Text -- ^ The prefix used
                -> L.Text -- ^ The command string, without a prefix
-               -> P.Sem r ()
-handleCommands handler msg prefix cmd = do
-  err <- P.runError $ do
-    (command, unparsedParams) <- P.fromEither $ mapLeft NF $ findCommand handler cmd
-    ctx <- P.note NoCtx =<< buildContext msg prefix command unparsedParams
-    P.fromEither . mapLeft (ERR ctx) =<< invokeCommand ctx (ctx ^. #command)
+               -> P.Sem r (Either CmdInvokeFailReason Context)
+handleCommands handler msg prefix cmd = P.runError $ do
+    (command, unparsedParams) <- P.fromEither $ mapLeft NotFound $ findCommand handler cmd
+    ctx <- P.note NoContext =<< buildContext msg prefix command unparsedParams
+    P.fromEither . mapLeft (CommandInvokeError ctx) =<< invokeCommand ctx (ctx ^. #command)
     pure ctx
-  case err of
-    Left (ERR ctx e) -> fire $ customEvt @"command-error" (ctx, e)
-    Left (NF path)   -> fire $ customEvt @"command-not-found" path
-    Left _           -> pure () -- ignore if no prefix or if context couldn't be built
-    Right ctx        -> do
-      cmdInvoke <- registerCounter "commands_invoked" [("name", S.unwords $ commandPath (ctx ^. #command))]
-      void $ addCounter 1 cmdInvoke
-      fire $ customEvt @"command-invoked" ctx
 
 
 -- | Run a command DSL, returning the constructed 'CommandHandler'